requirements.txt
# edge_gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic>=2.0,<3.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
pyyaml==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
aiofiles==23.2.1
jinja2==3.1.2
prometheus-client==0.19.0
structlog==23.2.0
httpx==0.25.2
pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.7
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.2
streamlit==1.28.1
jupyter==1.0.0
ipywidgets==8.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
mypy==1.6.1
pre-commit==3.5.0


# Rust Python bindings
maturin==1.9.1

# # Edge computing specific
# onnxruntime==1.16.3
# tensorrt==8.6.1
# openvino==2023.2.0
# https://github.com/pypa/pip/issues/6334

# Development tools
watchdog==3.0.0
rich==13.6.0
typer==0.9.0